requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
sqlmodel
python-dotenv